Custom UI backend data adapter (#124)
* adds boolean flag create_tables for shared database adapter that controls whether table creation should be tried on initialization or not. Necessary for integration tests and multi-instance setups, where we do not want to try creation multiple times due to timing issues.

* creates custom database adapter module for ui_backend without a singleton pattern. Inherit common behavior from shared data adapter.

* changes ui_backend components to explicitly require an initialized db, instead of falling back to singleton instance.

* creates UI specific table and model classes for ui_backend database adapter. Inherit common attributes from shared data adapter, such as schema version, create command, and keys.

* changes heartbeat monitor to use heartbeat_threshold from ui backend data adapter.

* changes integration tests to use ui_backend database adapter instead of shared one.

* removes all ui specific functionality from shared data module, moving them to the ui_backend database adapter instead.
