+++ services/data/postgres_async_db.py
@@ -18,9 +18,6 @@ from services.utils import DBConfiguration
 AIOPG_ECHO = os.environ.get("AIOPG_ECHO", 0) == "1"
 
 WAIT_TIME = 10
-# Heartbeat check interval. Add margin in case of client-server communication delays, before marking a heartbeat stale.
-HEARTBEAT_THRESHOLD = int(os.environ.get("HEARTBEAT_THRESHOLD", WAIT_TIME * 6))
-OLD_RUN_FAILURE_CUTOFF_TIME = int(os.environ.get("OLD_RUN_FAILURE_CUTOFF_TIME", 60 * 60 * 24 * 1000 * 14))  # default 2 weeks (in milliseconds)
 
 # Create database triggers automatically, disabled by default
 # Enable with env variable `DB_TRIGGER_CREATE=1`
@@ -58,7 +55,7 @@ class _AsyncPostgresDB(object):
         tables.append(self.metadata_table_postgres)
         self.tables = tables
 
-    async def _init(self, db_conf: DBConfiguration, create_triggers=DB_TRIGGER_CREATE):
+    async def _init(self, db_conf: DBConfiguration, create_triggers=DB_TRIGGER_CREATE, create_tables=True):
         # todo make poolsize min and max configurable as well as timeout
         # todo add retry and better error message
         retries = 3
@@ -77,7 +74,7 @@ class _AsyncPostgresDB(object):
                     await PostgresUtils.function_cleanup(self)
 
                 for table in self.tables:
-                    await table._init(create_triggers=create_triggers)
+                    await table._init(create_tables=create_tables, create_triggers=create_triggers)
 
                 self.logger.info(
                     "Connection established.\n"
@@ -110,45 +107,6 @@ class _AsyncPostgresDB(object):
                                                        step_name, task_name,
                                                        expanded=True)
         return task.body['task_id'], task.body['task_name']
-
-    # This function is used to verify 'data' object matches the same filters as
-    # 'AsyncPostgresTable.find_records' does. This is used with 'pg_notify' + Websocket
-    # events to make sure that specific subscriber receives filtered data correctly.
-    async def apply_filters_to_data(self, data, conditions: List[str] = None, values=[]) -> bool:
-        keys, vals, stm_vals = [], [], []
-        for k, v in data.items():
-            keys.append(k)
-            if k == "tags" or k == "system_tags":
-                # Handle JSON fields
-                vals.append(json.dumps(v))
-                stm_vals.append("%s::jsonb")
-            else:
-                vals.append(v)
-                stm_vals.append("%s")
-
-        # Prepend constructed data values before WHERE values
-        values = vals + values
-
-        select_sql = "SELECT * FROM (VALUES({values})) T({keys}) {where}".format(
-            values=", ".join(stm_vals),
-            keys=", ".join(map(lambda k: "\"{}\"".format(k), keys)),
-            where="WHERE {}".format(" AND ".join(
-                conditions)) if conditions else "",
-        )
-
-        try:
-            with (
-                await self.pool.cursor(
-                    cursor_factory=psycopg2.extras.DictCursor
-                )
-            ) as cur:
-                await cur.execute(select_sql, values)
-                records = await cur.fetchall()
-                cur.close()
-                return len(records) > 0
-        except:
-            self.logger.exception("Exception occured")
-            return False
 
 
 class AsyncPostgresDB(object):
@@ -189,8 +147,9 @@ class AsyncPostgresTable(object):
             raise NotImplementedError(
                 "need to specify table name and create command")
 
-    async def _init(self, create_triggers: bool):
-        await PostgresUtils.create_if_missing(self.db, self.table_name, self._command)
+    async def _init(self, create_tables: bool, create_triggers: bool):
+        if create_tables:
+            await PostgresUtils.create_if_missing(self.db, self.table_name, self._command)
         if create_triggers:
             self.db.logger.info(
                 "Create notify trigger for {table_name}\n   Keys: {keys}".format(
@@ -205,165 +164,41 @@ class AsyncPostgresTable(object):
             conditions.append("{} = %s".format(col_name))
             values.append(col_val)
 
-        response, *_ = await self.find_records(conditions=conditions, values=values, fetch_single=fetch_single,
-                                               order=ordering, limit=limit, expanded=expanded)
+        response, _ = await self.find_records(
+            conditions=conditions, values=values, fetch_single=fetch_single,
+            order=ordering, limit=limit, expanded=expanded
+        )
         return response
 
     async def find_records(self, conditions: List[str] = None, values=[], fetch_single=False,
-                           limit: int = 0, offset: int = 0, order: List[str] = None, groups: List[str] = None,
-                           group_limit: int = 10, expanded=False, enable_joins=False,
-                           postprocess: Callable[[DBResponse], DBResponse] = None,
-                           benchmark: bool = False, overwrite_select_from: str = None
-                           ) -> (DBResponse, DBPagination):
-        # Grouping not enabled
-        if groups is None or len(groups) == 0:
-            sql_template = """
-            SELECT * FROM (
-                SELECT
-                    {keys}
-                FROM {table_name}
-                {joins}
-            ) T
-            {where}
-            {order_by}
-            {limit}
-            {offset}
-            """
+                           limit: int = 0, offset: int = 0, order: List[str] = None, expanded=False,
+                           enable_joins=False) -> (DBResponse, DBPagination):
+        sql_template = """
+        SELECT * FROM (
+            SELECT
+                {keys}
+            FROM {table_name}
+            {joins}
+        ) T
+        {where}
+        {order_by}
+        {limit}
+        {offset}
+        """
 
-            select_sql = sql_template.format(
-                keys=",".join(
-                    self.select_columns + (self.join_columns if enable_joins and self.join_columns else [])),
-                table_name=overwrite_select_from if overwrite_select_from else self.table_name,
-                joins=" ".join(
-                    self.joins) if enable_joins and self.joins else "",
-                where="WHERE {}".format(" AND ".join(
-                    conditions)) if conditions else "",
-                order_by="ORDER BY {}".format(
-                    ", ".join(order)) if order else "",
-                limit="LIMIT {}".format(limit) if limit else "",
-                offset="OFFSET {}".format(offset) if offset else ""
-            ).strip()
-        else:  # Grouping enabled
-            # NOTE: we are performing a DISTINCT select on the group labels before the actual window function, to limit the set
-            # being queried. Without this restriction the query planner kept hitting the whole table contents, resulting in very slow queries.
+        select_sql = sql_template.format(
+            keys=",".join(
+                self.select_columns + (self.join_columns if enable_joins and self.join_columns else [])),
+            table_name=self.table_name,
+            joins=" ".join(self.joins) if enable_joins and self.joins is not None else "",
+            where="WHERE {}".format(" AND ".join(conditions)) if conditions else "",
+            order_by="ORDER BY {}".format(", ".join(order)) if order else "",
+            limit="LIMIT {}".format(limit) if limit else "",
+            offset="OFFSET {}".format(offset) if offset else ""
+        ).strip()
 
-            # Query for groups matching filters.
-            groups_sql_template = """
-            SELECT DISTINCT ON({group_by}) * FROM (
-                SELECT
-                    {keys}
-                FROM {table_name}
-                {joins}
-            ) T
-            {where}
-            ORDER BY {group_by} ASC NULLS LAST
-            {limit}
-            {offset}
-            """
-
-            groups_sql = groups_sql_template.format(
-                keys=",".join(
-                    self.select_columns + (self.join_columns if enable_joins and self.join_columns else [])),
-                table_name=self.table_name,
-                joins=" ".join(
-                    self.joins) if enable_joins and self.joins is not None else "",
-                where="WHERE {}".format(" AND ".join(
-                    conditions)) if conditions else "",
-                group_by=", ".join(groups),
-                limit="LIMIT {}".format(limit) if limit else "",
-                offset="OFFSET {}".format(offset) if offset else ""
-            ).strip()
-
-            group_results, _ = await self.execute_sql(select_sql=groups_sql, values=values, fetch_single=fetch_single,
-                                                      expanded=expanded, limit=limit, offset=offset)
-            if len(group_results.body) == 0:
-                # Return early if no groups match the query.
-                return group_results, None, None
-
-            # construct the group_where clause.
-            group_label_selects = []
-            for group in groups:
-                _group_values = []
-                for row in group_results.body:
-                    _group_values.append("\'{}\'".format(row[group.strip("\"")]))
-                if len(_group_values) > 0:
-                    _clause = "{group} IN ({values})".format(group=group, values=", ".join(_group_values))
-                    group_label_selects.append(_clause)
-
-            # Query for group content. Restricted by groups received from previous query.
-            sql_template = """
-            SELECT * FROM (
-                SELECT
-                    *, ROW_NUMBER() OVER(PARTITION BY {group_by} {order_by})
-                FROM (
-                    SELECT
-                        {keys}
-                    FROM {table_name}
-                    {joins}
-                ) T
-                {where}
-            ) G
-            {group_where}
-            """
-
-            select_sql = sql_template.format(
-                keys=",".join(
-                    self.select_columns + (self.join_columns if enable_joins and self.join_columns else [])),
-                table_name=overwrite_select_from if overwrite_select_from else self.table_name,
-                joins=" ".join(
-                    self.joins) if enable_joins and self.joins is not None else "",
-                where="WHERE {}".format(" AND ".join(
-                    conditions)) if conditions else "",
-                group_by=", ".join(groups),
-                order_by="ORDER BY {}".format(
-                    ", ".join(order)) if order else "",
-                group_where="""
-                    WHERE {group_limit} {group_selects}
-                """.format(
-                    group_limit="row_number <= {} AND ".format(group_limit) if group_limit else "",
-                    group_selects=" AND ".join(group_label_selects)
-                )
-            ).strip()
-
-        # Run benchmarking on query if requested
-        benchmark_results = None
-        if benchmark:
-            benchmark_results = await self.benchmark_sql(
-                select_sql=select_sql, values=values, fetch_single=fetch_single,
-                expanded=expanded, limit=limit, offset=offset
-            )
-
-        result, pagination = await self.execute_sql(select_sql=select_sql, values=values, fetch_single=fetch_single,
-                                                    expanded=expanded, limit=limit, offset=offset)
-        # Modify the response after the fetch has been executed
-        if postprocess is not None:
-            if iscoroutinefunction(postprocess):
-                result = await postprocess(result)
-            else:
-                result = postprocess(result)
-        return result, pagination, benchmark_results
-
-    async def benchmark_sql(self, select_sql: str, values=[], fetch_single=False,
-                            expanded=False, limit: int = 0, offset: int = 0):
-        "Benchmark and log a given SQL query with EXPLAIN ANALYZE"
-        try:
-            with (
-                await self.db.pool.cursor(
-                    cursor_factory=psycopg2.extras.DictCursor
-                )
-            ) as cur:
-                # Run EXPLAIN ANALYZE on query and log the results.
-                benchmark_sql = "EXPLAIN ANALYZE {}".format(select_sql)
-                await cur.execute(benchmark_sql, values)
-
-                records = await cur.fetchall()
-                rows = []
-                for record in records:
-                    rows.append(record[0])
-                return "\n".join(rows)
-        except (Exception, psycopg2.DatabaseError):
-            self.db.logger.exception("Query Benchmarking failed")
-            return None
+        return await self.execute_sql(select_sql=select_sql, values=values, fetch_single=fetch_single,
+                                      expanded=expanded, limit=limit, offset=offset)
 
     async def execute_sql(self, select_sql: str, values=[], fetch_single=False,
                           expanded=False, limit: int = 0, offset: int = 0) -> (DBResponse, DBPagination):
@@ -400,28 +235,6 @@ class AsyncPostgresTable(object):
         except (Exception, psycopg2.DatabaseError) as error:
             self.db.logger.exception("Exception occured")
             return aiopg_exception_handling(error), None
-
-    async def get_tags(self):
-        sql_template = "SELECT DISTINCT tag FROM (SELECT JSONB_ARRAY_ELEMENTS(tags||system_tags) AS tag FROM {table_name}) AS t"
-        select_sql = sql_template.format(table_name=self.table_name)
-
-        try:
-            with (
-                await self.db.pool.cursor(
-                    cursor_factory=psycopg2.extras.DictCursor
-                )
-            ) as cur:
-                await cur.execute(select_sql)
-
-                tags = []
-                records = await cur.fetchall()
-                for record in records:
-                    tags += record
-                cur.close()
-                return DBResponse(response_code=200, body=tags)
-        except (Exception, psycopg2.DatabaseError) as error:
-            self.db.logger.exception("Exception occured")
-            return aiopg_exception_handling(error)
 
     async def create_record(self, record_dict):
         # note: need to maintain order
@@ -623,7 +436,6 @@ class AsyncFlowTablePostgres(AsyncPostgresTable):
     primary_keys = ["flow_id"]
     trigger_keys = primary_keys
     select_columns = keys
-    join_columns = []
     _command = """
     CREATE TABLE {0} (
         flow_id VARCHAR(255) PRIMARY KEY,
@@ -664,101 +476,7 @@ class AsyncRunTablePostgres(AsyncPostgresTable):
             "user_name", "ts_epoch", "last_heartbeat_ts", "tags", "system_tags"]
     primary_keys = ["flow_id", "run_number"]
     trigger_keys = primary_keys
-    joins = [
-        """
-        LEFT JOIN (
-            SELECT
-                artifacts.flow_id, artifacts.run_number, artifacts.step_name,
-                artifacts.task_id, artifacts.attempt_id, artifacts.ts_epoch,
-                attempt_ok.value::boolean as attempt_ok
-            FROM {artifact_table} as artifacts
-            LEFT JOIN {metadata_table} as attempt_ok ON (
-                artifacts.flow_id = attempt_ok.flow_id AND
-                artifacts.run_number = attempt_ok.run_number AND
-                artifacts.task_id = attempt_ok.task_id AND
-                artifacts.step_name = attempt_ok.step_name AND
-                attempt_ok.field_name = 'attempt_ok' AND
-                attempt_ok.tags ? ('attempt_id:' || artifacts.attempt_id)
-            )
-            WHERE artifacts.name = '_task_ok' AND artifacts.step_name = 'end'
-        ) AS artifacts ON (
-            {table_name}.flow_id = artifacts.flow_id AND
-            {table_name}.run_number = artifacts.run_number
-        )
-        """.format(
-            table_name=table_name,
-            metadata_table="metadata_v3",
-            artifact_table="artifact_v3"
-        ),
-    ]
-    # User should be considered NULL when 'user:*' tag is missing
-    # This is usually the case with AWS Step Functions
-    select_columns = ["runs_v3.{0} AS {0}".format(k) for k in keys] \
-        + ["""
-            (CASE
-                WHEN system_tags ? ('user:' || user_name)
-                THEN user_name
-                ELSE NULL
-            END) AS user"""] \
-        + ["""
-            COALESCE({table_name}.run_id, {table_name}.run_number::text) AS run
-            """.format(table_name=table_name)]
-    join_columns = [
-        """
-        (CASE
-            WHEN artifacts.ts_epoch IS NOT NULL
-            THEN artifacts.ts_epoch
-            WHEN {table_name}.last_heartbeat_ts IS NOT NULL
-            AND @(extract(epoch from now())-{table_name}.last_heartbeat_ts)>{heartbeat_threshold}
-            THEN {table_name}.last_heartbeat_ts*1000
-            WHEN {table_name}.last_heartbeat_ts IS NULL
-            AND @(extract(epoch from now())*1000-{table_name}.ts_epoch)>{cutoff}
-            THEN {table_name}.ts_epoch + {cutoff}
-            ELSE NULL
-        END) AS finished_at
-        """.format(
-            table_name=table_name,
-            heartbeat_threshold=HEARTBEAT_THRESHOLD,
-            cutoff=OLD_RUN_FAILURE_CUTOFF_TIME
-        ),
-        """
-        (CASE
-            WHEN artifacts.attempt_ok IS TRUE
-            THEN 'completed'
-            WHEN artifacts.attempt_ok IS FALSE
-            THEN 'failed'
-            WHEN artifacts.ts_epoch IS NOT NULL
-            THEN 'completed'
-            WHEN {table_name}.last_heartbeat_ts IS NOT NULL
-            AND @(extract(epoch from now())-{table_name}.last_heartbeat_ts)>{heartbeat_threshold}
-            THEN 'failed'
-            WHEN {table_name}.last_heartbeat_ts IS NULL
-            AND @(extract(epoch from now())*1000-{table_name}.ts_epoch)>{cutoff}
-            THEN 'failed'
-            ELSE 'running'
-        END) AS status
-        """.format(
-            table_name=table_name,
-            heartbeat_threshold=HEARTBEAT_THRESHOLD,
-            cutoff=OLD_RUN_FAILURE_CUTOFF_TIME
-        ),
-        """
-        (CASE
-            WHEN artifacts.ts_epoch IS NULL AND {table_name}.last_heartbeat_ts IS NOT NULL
-            THEN {table_name}.last_heartbeat_ts*1000-{table_name}.ts_epoch
-            WHEN artifacts.ts_epoch IS NOT NULL
-            THEN artifacts.ts_epoch - {table_name}.ts_epoch
-            WHEN {table_name}.last_heartbeat_ts IS NULL
-            AND @(extract(epoch from now())::bigint*1000-{table_name}.ts_epoch)>{cutoff}
-            THEN {cutoff}
-            ELSE @(extract(epoch from now())::bigint*1000-{table_name}.ts_epoch)
-        END) AS duration
-        """.format(
-            table_name=table_name,
-            heartbeat_threshold=HEARTBEAT_THRESHOLD,
-            cutoff=OLD_RUN_FAILURE_CUTOFF_TIME
-        )
-    ]
+    select_columns = keys
     flow_table_name = AsyncFlowTablePostgres.table_name
     _command = """
     CREATE TABLE {0} (
@@ -824,51 +542,6 @@ class AsyncStepTablePostgres(AsyncPostgresTable):
     trigger_keys = primary_keys
     select_columns = keys
     run_table_name = AsyncRunTablePostgres.table_name
-    task_table_name = "tasks_v3"
-    artifact_table_name = "artifact_v3"
-    joins = [
-        """
-        LEFT JOIN LATERAL (
-            SELECT last_heartbeat_ts as heartbeat_ts
-            FROM {task_table}
-            WHERE {table_name}.flow_id={task_table}.flow_id
-            AND {table_name}.run_number={task_table}.run_number
-            AND {table_name}.step_name={task_table}.step_name
-            ORDER BY last_heartbeat_ts DESC
-            LIMIT 1
-        ) AS latest_task_hb ON true
-        """.format(
-            table_name=table_name,
-            task_table=task_table_name
-        ),
-        """
-        LEFT JOIN LATERAL (
-            SELECT ts_epoch as ts_epoch
-            FROM {artifact_table}
-            WHERE {table_name}.flow_id={artifact_table}.flow_id
-            AND {table_name}.run_number={artifact_table}.run_number
-            AND {table_name}.step_name={artifact_table}.step_name
-            AND {artifact_table}.name = '_task_ok'
-            ORDER BY
-                ts_epoch DESC
-            LIMIT 1
-        ) AS latest_task_ok ON true
-        """.format(
-            table_name=table_name,
-            artifact_table=artifact_table_name
-        )
-    ]
-    select_columns = ["steps_v3.{0} AS {0}".format(k) for k in keys]
-    join_columns = [
-        """
-        GREATEST(
-            latest_task_ok.ts_epoch,
-            latest_task_hb.heartbeat_ts*1000
-        ) - {table_name}.ts_epoch as duration
-        """.format(
-            table_name=table_name
-        )
-    ]
     _command = """
     CREATE TABLE {0} (
         flow_id VARCHAR(255) NOT NULL,
@@ -925,134 +598,7 @@ class AsyncTaskTablePostgres(AsyncPostgresTable):
             "task_name", "user_name", "ts_epoch", "last_heartbeat_ts", "tags", "system_tags"]
     primary_keys = ["flow_id", "run_number", "step_name", "task_id"]
     trigger_keys = primary_keys
-    # NOTE: There is a lot of unfortunate backwards compatibility for cases where task metadata, or artifacts
-    # have not been stored correctly.
-    # NOTE: tasks_v3 table does not have a column for 'attempt_id', instead this is added before the join
-    # with a subquery in the FROM.
-    # NOTE: when using these joins, we _must_ clean up the results with a WHERE that discards attempts with
-    # nothing joined, otherwise we end up with ghost attempts for the task.
-    joins = [
-        """
-        LEFT JOIN {metadata_table} as start ON (
-            {table_name}.flow_id = start.flow_id AND
-            {table_name}.run_number = start.run_number AND
-            {table_name}.step_name = start.step_name AND
-            {table_name}.task_id = start.task_id AND
-            start.field_name = 'attempt' AND
-            {table_name}.attempt_id = start.value::int
-        )
-        LEFT JOIN {metadata_table} as done ON (
-            {table_name}.flow_id = done.flow_id AND
-            {table_name}.run_number = done.run_number AND
-            {table_name}.step_name = done.step_name AND
-            {table_name}.task_id = done.task_id AND
-            done.field_name = 'attempt-done' AND
-            {table_name}.attempt_id = done.value::int
-        )
-        LEFT JOIN {metadata_table} as next_attempt_start ON (
-            {table_name}.flow_id = next_attempt_start.flow_id AND
-            {table_name}.run_number = next_attempt_start.run_number AND
-            {table_name}.step_name = next_attempt_start.step_name AND
-            {table_name}.task_id = next_attempt_start.task_id AND
-            next_attempt_start.field_name = 'attempt' AND
-            ({table_name}.attempt_id + 1) = next_attempt_start.value::int
-        )
-        LEFT JOIN {metadata_table} as attempt_ok ON (
-            {table_name}.flow_id = attempt_ok.flow_id AND
-            {table_name}.run_number = attempt_ok.run_number AND
-            {table_name}.step_name = attempt_ok.step_name AND
-            {table_name}.task_id = attempt_ok.task_id AND
-            attempt_ok.field_name = 'attempt_ok' AND
-            attempt_ok.tags ? ('attempt_id:' || {table_name}.attempt_id)
-        )
-        LEFT JOIN {artifact_table} as foreach_stack ON (
-            {table_name}.flow_id = foreach_stack.flow_id AND
-            {table_name}.run_number = foreach_stack.run_number AND
-            {table_name}.step_name = foreach_stack.step_name AND
-            {table_name}.task_id = foreach_stack.task_id AND
-            foreach_stack.name = '_foreach_stack' AND
-            {table_name}.attempt_id = foreach_stack.attempt_id
-        )
-        LEFT JOIN {artifact_table} as task_ok ON (
-            {table_name}.flow_id = task_ok.flow_id AND
-            {table_name}.run_number = task_ok.run_number AND
-            {table_name}.step_name = task_ok.step_name AND
-            {table_name}.task_id = task_ok.task_id AND
-            task_ok.name = '_task_ok' AND
-            {table_name}.attempt_id = task_ok.attempt_id
-        )
-        """.format(
-            table_name=table_name,
-            metadata_table="metadata_v3",
-            artifact_table="artifact_v3"
-        ),
-    ]
-    select_columns = ["tasks_v3.{0} AS {0}".format(k) for k in keys]
-    join_columns = [
-        "{table_name}.attempt_id as attempt_id".format(table_name=table_name),
-        "start.ts_epoch as started_at",
-        """
-        (CASE
-        WHEN {finished_at_column} IS NULL
-            AND {table_name}.last_heartbeat_ts IS NOT NULL
-            AND @(extract(epoch from now())-{table_name}.last_heartbeat_ts)>{heartbeat_threshold}
-        THEN {table_name}.last_heartbeat_ts*1000
-        ELSE {finished_at_column}
-        END) as finished_at
-        """.format(
-            table_name=table_name,
-            heartbeat_threshold=HEARTBEAT_THRESHOLD,
-            finished_at_column="COALESCE(attempt_ok.ts_epoch, done.ts_epoch, task_ok.ts_epoch, next_attempt_start.ts_epoch)"
-        ),
-        "attempt_ok.value::boolean as attempt_ok",
-        # If 'attempt_ok' is present, we can leave task_ok NULL since
-        #   that is used to fetch the artifact value from remote location.
-        # This process is performed at TaskRefiner (data_refiner.py)
-        """
-        (CASE
-            WHEN attempt_ok.ts_epoch IS NOT NULL
-            THEN NULL
-            ELSE task_ok.location
-        END) as task_ok
-        """,
-        """
-        (CASE
-            WHEN attempt_ok.value::boolean IS TRUE
-            THEN 'completed'
-            WHEN attempt_ok.value::boolean IS FALSE
-            THEN 'failed'
-            WHEN COALESCE(done.ts_epoch, task_ok.ts_epoch) IS NOT NULL
-                AND attempt_ok IS NULL
-            THEN 'unknown'
-            WHEN COALESCE(attempt_ok.ts_epoch, done.ts_epoch, task_ok.ts_epoch) IS NOT NULL
-            THEN 'completed'
-            WHEN next_attempt_start.ts_epoch IS NOT NULL
-            THEN 'failed'
-            WHEN {finished_at_column} IS NULL
-                AND {table_name}.last_heartbeat_ts IS NOT NULL
-                AND @(extract(epoch from now())-{table_name}.last_heartbeat_ts)>{heartbeat_threshold}
-            THEN 'failed'
-            ELSE 'running'
-        END) AS status
-        """.format(
-            table_name=table_name,
-            heartbeat_threshold=HEARTBEAT_THRESHOLD,
-            finished_at_column="COALESCE(attempt_ok.ts_epoch, done.ts_epoch, task_ok.ts_epoch)"
-        ),
-        """
-        (CASE
-            WHEN {finished_at_column} IS NULL AND {table_name}.last_heartbeat_ts IS NOT NULL
-            THEN {table_name}.last_heartbeat_ts*1000-COALESCE(start.ts_epoch, {table_name}.ts_epoch)
-            WHEN {finished_at_column} IS NOT NULL
-            THEN {finished_at_column} - COALESCE(start.ts_epoch, {table_name}.ts_epoch)
-            ELSE NULL
-        END) AS duration
-        """.format(
-            table_name=table_name,
-            finished_at_column="COALESCE(attempt_ok.ts_epoch, done.ts_epoch, task_ok.ts_epoch, next_attempt_start.ts_epoch)"
-        ),
-        "foreach_stack.location as foreach_stack"
-    ]
+    select_columns = keys
     step_table_name = AsyncStepTablePostgres.table_name
     _command = """
     CREATE TABLE {0} (
@@ -1128,23 +674,6 @@ class AsyncTaskTablePostgres(AsyncPostgresTable):
 
         return DBResponse(response_code=result.response_code,
                           body=json.dumps(body))
-
-    async def find_records(self, conditions: List[str] = None, values=[], fetch_single=False,
-                           limit: int = 0, offset: int = 0, order: List[str] = None, groups: List[str] = None,
-                           group_limit: int = 10, expanded=False, enable_joins=False,
-                           postprocess: Callable[[DBResponse], DBResponse] = None,
-                           benchmark: bool = False, overwrite_select_from: str = None
-                           ) -> (DBResponse, DBPagination):
-        if enable_joins:
-            overwrite_select_from = "(SELECT *, UNNEST('{0, 1, 2, 3, 4}'::int[]) as attempt_id FROM tasks_v3) as tasks_v3"
-            conditions.append("NOT (attempt_id > 0 AND started_at IS NULL AND task_ok IS NULL)")
-        return await super().find_records(
-            conditions, values, fetch_single,
-            limit, offset, order,
-            groups, group_limit, expanded,
-            enable_joins, postprocess, benchmark,
-            overwrite_select_from
-        )
 
 
 class AsyncMetadataTablePostgres(AsyncPostgresTable):
