+++ services/ui_backend_service/api/artifact.py
@@ -1,13 +1,12 @@
 import json
 from aiohttp import web
-from services.data.postgres_async_db import AsyncPostgresDB
 from services.data.db_utils import DBResponse, _filter_artifacts_by_attempt_id_for_tasks, translate_run_key, translate_task_key
 from services.utils import handle_exceptions
 from .utils import find_records
 
 
 class ArtificatsApi(object):
-    def __init__(self, app, db=AsyncPostgresDB.get_instance()):
+    def __init__(self, app, db):
         self.db = db
         app.router.add_route(
             "GET",
