+++ services/ui_backend_service/api/heartbeat_monitor.py
@@ -2,9 +2,8 @@ import asyncio
 import datetime
 from typing import Dict
 from pyee import AsyncIOEventEmitter
-from services.data.postgres_async_db import AsyncPostgresDB
 from services.data.db_utils import translate_run_key, translate_task_key
-from services.data.postgres_async_db import HEARTBEAT_THRESHOLD
+from ..data.db.tables.base import HEARTBEAT_THRESHOLD
 from .notify import resource_list
 from ..data.refiner import TaskRefiner
 
@@ -14,7 +13,7 @@ HEARTBEAT_INTERVAL = HEARTBEAT_THRESHOLD + 10
 
 
 class HeartbeatMonitor(object):
-    def __init__(self, event_name, event_emitter=None, db=AsyncPostgresDB.get_instance(), cache=None):
+    def __init__(self, event_name, db, event_emitter=None, cache=None):
         self.watched = {}
         # Handle HB Events
         self.event_emitter = event_emitter or AsyncIOEventEmitter()
@@ -76,8 +75,8 @@ class RunHeartbeatMonitor(HeartbeatMonitor):
         # Init the abstract class
         super().__init__(
             event_name="run-heartbeat",
-            event_emitter=event_emitter,
-            db=db
+            db=db,
+            event_emitter=event_emitter
         )
         # Table for data fetching for load_and_broadcast and add_to_watch
         self._run_table = self.db.run_table_postgres
@@ -141,8 +140,8 @@ class TaskHeartbeatMonitor(HeartbeatMonitor):
         # Init the abstract class
         super().__init__(
             event_name="task-heartbeat",
-            event_emitter=event_emitter,
             db=db,
+            event_emitter=event_emitter,
             cache=cache
         )
         # Table for data fetching for load_and_broadcast and add_to_watch
