+++ services/ui_backend_service/api/notify.py
@@ -1,13 +1,12 @@
 import json
 import asyncio
 from typing import Dict, List
-from services.data.postgres_async_db import AsyncPostgresDB
 from services.utils import logging
 from pyee import AsyncIOEventEmitter
 
 
 class ListenNotify(object):
-    def __init__(self, app, event_emitter=None, db=AsyncPostgresDB.get_instance()):
+    def __init__(self, app, db, event_emitter=None):
         self.event_emitter = event_emitter or AsyncIOEventEmitter()
         self.db = db
         self.logger = logging.getLogger("ListenNotify")
