+++ services/ui_backend_service/api/tag.py
@@ -1,9 +1,8 @@
-from services.data.postgres_async_db import AsyncPostgresDB
 from services.utils import handle_exceptions, web_response
 
 
 class TagApi(object):
-    def __init__(self, app, db=AsyncPostgresDB.get_instance()):
+    def __init__(self, app, db):
         self.db = db
         app.router.add_route("GET", "/tags", self.get_all_tags)
         self._async_table = self.db.run_table_postgres
