+++ services/ui_backend_service/data/db/__init__.py
@@ -1,0 +1,1 @@
+from .postgres_async_db import AsyncPostgresDB
