+++ services/ui_backend_service/data/db/models/base_row.py
@@ -1,0 +1,3 @@
+class BaseRow(object):
+    def serialize(self):
+        raise NotImplementedError("Row model needs to define a serialize function")
