+++ services/ui_backend_service/data/db/tables/flow.py
@@ -1,0 +1,14 @@
+from .base import AsyncPostgresTable
+from ..models import FlowRow
+# use schema constants from the .data module to keep things consistent
+from services.data.postgres_async_db import AsyncFlowTablePostgres as MetadataFlowTable
+
+
+class AsyncFlowTablePostgres(AsyncPostgresTable):
+    table_name = MetadataFlowTable.table_name
+    keys = MetadataFlowTable.keys
+    primary_keys = MetadataFlowTable.primary_keys
+    trigger_keys = MetadataFlowTable.trigger_keys
+    select_columns = keys
+    _command = MetadataFlowTable._command
+    _row_type = FlowRow
